Remove python 2 support
